@sxllwx 2022-10-26
> I'm not convinced we should deep copy at all. But definitely it shouldn't go here, there are other callers of this code path than just the watch cache.

Thank you for your suggestion ~ 😄

I find a more suitable place to add the relevant logic.