@lavalamp 2022-10-28
This still double-allocates for some amount of time, it's too late.

I'm actually not even convinced that deep copy is enough, are e.g. strings copied or are they still references to the original deserialized data?

A correct fix is going to be very invasive, which is why I'm questioning that the benefit would be high enough.