@sxllwx 2022-10-29
```bash
(⎈ |test:kube-system)➜  meta git:(ftr/extract_listobject_use_copy) go test -v --bench BenchmarkExtractList --benchmem help_test.go help.go
goos: darwin
goarch: amd64
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkExtractList
BenchmarkExtractList-12                     4009            473747 ns/op          513492 B/op       3006 allocs/op
BenchmarkExtractListWithAlloc
BenchmarkExtractListWithAlloc-12             975           1084318 ns/op          514468 B/op       3009 allocs/op
PASS
ok      command-line-arguments  4.052s
```

Here is my local benchmark data for your reference. At present, `ExtractListWithAllloc` is already equivalent to `ExtractList` in terms of memory allocation.