@sxllwx 2022-10-29
Some clarifications are necessary here:

Take `corev1.ConfigMapList` as an example (for the sake of convenience, I directly copied the relevant part of the code to prevent jumping around):

```golang
type ConfigMapList struct {
         ....
	Items []ConfigMap
}

type ConfigMap struct {
	metav1.TypeMeta `json:",inline"`
	metav1.ObjectMeta `json:"metadata,omitempty" protobuf:"bytes,1,opt,name=metadata"`
        ....
	Data map[string]string `json:"data,omitempty" protobuf:"bytes,2,rep,name=data"`
}

func (in *ConfigMap) DeepCopyObject() runtime.Object {
       ...
}
```

Items elements are all ConfigMap, ConfigMap does not implement runtime.Object, (the receiver of the DeepCopyObject method is *ConfigMap), so we need to use *ConfigMap.

The return value of our `ExtractList` is `[]runtime.Object`, so we have the following available:

1. Get the pointer of the ConfigMap.Items[i] element directly (the current implementation), but this will cause the Slice of ConfigMapList.Items to not be released

2. DeepCopy (Deprecated): Causes double memory consumption, ps: Not only will golang-basic type [int, string,] field values ​​(Kind, APIVersion, Name, UID, ...) be copied, when encountering To Data, or Labels, Annotations field type is Map[string]string will also be copied. 

3. Alloc a new ConfigMap, and assign a value to it. Execute code similar to the following:

`ret := configMapList.Items[0]`

Only the fields of the basic type (int, string... mentioned above) will be copied here, and the Data field will point to configMapList.Items[0].Data. When go-runtime executes the gc-routine, if configMapList.Items[0] has no related references. will be released. But since Data is referenced by ret.Data at the same time. This memory will not be freed.

> Here is a very strange question, why do method 1 and method 3 have the same number of allocations in the benchmark results?

This is because the following code exists in ExtractList:

````golang
func ExtractList(obj runtime.Object) ([]runtime.Object, error) {
	....
	for i := range list {
		raw := items.Index(i)
                
		switch item := raw.Interface().(type) {
		....
       ....
````

`items := raw.Interface().(type)` The code here has actually done an Alloc Object && Set to the ConfigMap

However, ExtractListWithAlloc determines the type by comparing raw.Type. When it is determined that the current Object is a ConfigMap, Alloc creates a new Configmap and assigns it a value. So it can not only guarantee the number of memory allocations, but also avoid directly referencing the Items list.