@liggitt 2022-10-29
> The map and slice in the newly generated object will share the same memory space with the elements in the list. Only the basic types of golang fields such as int and string will allocate memory.

non-pointer struct fields (like ObjectMeta, Spec, and Status) will also be allocated. I don't think ConfigMap is a good example to benchmark… deeply nested high-cardinality objects with struct fields like Pod and Node are a better test for the cost of a shallow copy