@sxllwx 2022-10-30
/ping @lavalamp  @liggitt 

I have added performance tests related to PodList and the results are as follows:

```bash
goos: linux
goarch: amd64
pkg: k8s.io/client-go/tools/cache
cpu: AMD EPYC 7K62 48-Core Processor
BenchmarkExtractPodList
BenchmarkExtractPodList-16             	     146	   8162330 ns/op	10031856 B/op	   31503 allocs/op
BenchmarkExtractPodListWithAlloc
BenchmarkExtractPodListWithAlloc-16    	     135	   8760651 ns/op	11056178 B/op	   32503 allocs/op
````

> Performance meets expectations 😄

In addition, a unit test (TestReflectorListExtract) has been added to prove the correctness of this modification. PTAL, thx~