@liggitt 2022-10-31
the reflect optimization is a good find... that appears to cover the added alloc cost of the shallow copy, which is lucky

I split apart your commit into steps that make it easier to see the before / after comparison and pushed to https://github.com/liggitt/kubernetes/commits/ftr/extract_listobject_use_copy if you're interested

Structuring the benchmarks so the test names stay the same and the performance changes makes it easy to use `benchcmp` to compare before/after results. Also, I moved the construction of the dummy list data outside the benchmark, so the benchmark is just checking the performance of extract list.

Results:

```
# benchmark with no changes
go test k8s.io/client-go/tools/cache -run NONE -bench Extract -benchmem > master.txt
BenchmarkExtractPodList-12          2324            528517 ns/op         1040394 B/op       1002 allocs/op

# benchmark with reflect change only
go test k8s.io/client-go/tools/cache -run NONE -bench Extract -benchmem > reflect.txt
BenchmarkExtractPodList-12         10000            104362 ns/op           16392 B/op          2 allocs/op

# benchmark with extractlist changes
go test k8s.io/client-go/tools/cache -run NONE -bench Extract -benchmem > allocs.txt
BenchmarkExtractPodList-12          1402           1050354 ns/op         1040395 B/op       1002 allocs/op
```

compare master to reflect change:
```
benchcmp master.txt reflect.txt 
benchmark                      old ns/op     new ns/op     delta
BenchmarkExtractPodList-12     528517        104362        -80.25%

benchmark                      old allocs     new allocs     delta
BenchmarkExtractPodList-12     1002           2              -99.80%

benchmark                      old bytes     new bytes     delta
BenchmarkExtractPodList-12     1040394       16392         -98.42%
```

compare reflect change to reflect+allocs change:
```
benchcmp reflect.txt allocs.txt 
benchmark                      old ns/op     new ns/op     delta
BenchmarkExtractPodList-12     104362        1050354       +906.45%

benchmark                      old allocs     new allocs     delta
BenchmarkExtractPodList-12     2              1002           +50000.00%

benchmark                      old bytes     new bytes     delta
BenchmarkExtractPodList-12     16392         1040395       +6246.97%
```

compare master to reflect+allocs change:
```
benchcmp master.txt allocs.txt 
benchmark                      old ns/op     new ns/op     delta
BenchmarkExtractPodList-12     528517        1050354       +98.74%

benchmark                      old allocs     new allocs     delta
BenchmarkExtractPodList-12     1002           1002           +0.00%

benchmark                      old bytes     new bytes     delta
BenchmarkExtractPodList-12     1040394       1040395       +0.00%
```

It looks like the reflect fix perfectly compensates for the additional allocs incurred, which is super lucky. The execution time is longer with the allocs change (.5ms → 1ms for a list of 1000 pods), but that may be worth it.