@sxllwx 2022-11-03
Update: 

Results from the latest benchmarks:

`apimachinery/pkg/api/meta`:

```golang
benchmark                                old ns/op     new ns/op     delta
BenchmarkExtractSampleList-16            235826        64968         -72.45%
BenchmarkExtractFooList-16               211822        314864        +48.65%
BenchmarkExtractRawExtensionList-16      83677         87361         +4.40%
BenchmarkEachSampleListItem-16           40323         319345        +691.97%
BenchmarkEachFooListItem-16              35270         302871        +758.72%
BenchmarkEachRawExtensionListItem-16     18588         26386         +41.95%

benchmark                                old allocs     new allocs     delta
BenchmarkExtractSampleList-16            1002           2              -99.80%
BenchmarkExtractFooList-16               1002           1002           +0.00%
BenchmarkExtractRawExtensionList-16      1002           1002           +0.00%
BenchmarkEachSampleListItem-16           1              1              +0.00%
BenchmarkEachFooListItem-16              1              1              +0.00%
BenchmarkEachRawExtensionListItem-16     1              1              +0.00%

benchmark                                old bytes     new bytes     delta
BenchmarkExtractSampleList-16            304395        16392         -94.61%
BenchmarkExtractFooList-16               304393        304394        +0.00%
BenchmarkExtractRawExtensionList-16      64392         64392         +0.00%
BenchmarkEachSampleListItem-16           8             8             +0.00%
BenchmarkEachFooListItem-16              8             8             +0.00%
BenchmarkEachRawExtensionListItem-16     8             8             +0.00%
```

`k8s.io/client-go/tools/cache`【no shallow copy】:

```golang
/tmp/cache »  benchcmp master reflect 
benchmark                               old ns/op     new ns/op     delta
BenchmarkExtractPodList-16              446505        8113          -98.18%
BenchmarkEachPodListItem-16             30623         32679         +6.71%
BenchmarkReflectorList-16               9013419       1195816       -86.73%

benchmark                               old allocs     new allocs     delta
BenchmarkExtractPodList-16              1002           2              -99.80%
BenchmarkEachPodListItem-16             1              1              +0.00%
BenchmarkReflectorList-16               44477          4348           -90.22%

benchmark                               old bytes     new bytes     delta
BenchmarkExtractPodList-16              1040392       1800          -99.83%
BenchmarkEachPodListItem-16             8             8             +0.00%
BenchmarkReflectorList-16               12110448      1201971       -90.07%
```

`k8s.io/client-go/tools/cache` 【shallow copy】:

```
benchmark                               old ns/op     new ns/op     delta
BenchmarkExtractPodList-16              446505        58443         -86.91%
BenchmarkEachPodListItem-16             30623         83883         +173.92%
BenchmarkReflectorList-16               9013419       1178825       -86.92%
BenchmarkIndexer-16                     309           304           -1.78%

benchmark                               old allocs     new allocs     delta
BenchmarkExtractPodList-16              1002           102            -89.82%
BenchmarkEachPodListItem-16             1              101            +10000.00%
BenchmarkReflectorList-16               44477          4348           -90.22%
BenchmarkIndexer-16                     3              3              +0.00%

benchmark                               old bytes     new bytes     delta
BenchmarkExtractPodList-16              1040392       104200        -89.98%
BenchmarkEachPodListItem-16             8             102408        +1280000.00%
BenchmarkReflectorList-16               12110448      1201970       -90.07%
BenchmarkIndexer-16                     48            48            +0.00%
```