@sxllwx 2022-11-05
Update(ready for review):

Compare master and (optimize reflect only):

> pkg: k8s.io/apimachinery/pkg/api/meta

```golang
benchcmp   master reflect                                                                                                        ubuntu@www
benchmark                                        old ns/op     new ns/op     delta
BenchmarkExtractListItem/FooList-16              183374        285066        +55.46%
BenchmarkExtractListItem/SampleList-16           214082        63593         -70.30%
BenchmarkExtractListItem/RawExtensionList-16     80363         87624         +9.04%
BenchmarkEachListItem/FooList-16                 34429         36168         +5.05%
BenchmarkEachListItem/SampleList-16              39539         39378         -0.41%
BenchmarkEachListItem/RawExtensionList-16        18509         19106         +3.23%

benchmark                                        old allocs     new allocs     delta
BenchmarkExtractListItem/FooList-16              1002           1002           +0.00%
BenchmarkExtractListItem/SampleList-16           1002           2              -99.80%
BenchmarkExtractListItem/RawExtensionList-16     1002           1002           +0.00%
BenchmarkEachListItem/FooList-16                 1              1              +0.00%
BenchmarkEachListItem/SampleList-16              1              1              +0.00%
BenchmarkEachListItem/RawExtensionList-16        1              1              +0.00%

benchmark                                        old bytes     new bytes     delta
BenchmarkExtractListItem/FooList-16              304395        304394        -0.00%
BenchmarkExtractListItem/SampleList-16           304392        16392         -94.61%
BenchmarkExtractListItem/RawExtensionList-16     64392         64392         +0.00%
BenchmarkEachListItem/FooList-16                 8             8             +0.00%
BenchmarkEachListItem/SampleList-16              8             8             +0.00%
BenchmarkEachListItem/RawExtensionList-16        8             8             +0.00%
```

> pkg: k8s.io/client-go/tools/cache

```golang
benchcmp   master reflect                                                                                                       ubuntu@www
benchmark                                   old ns/op     new ns/op     delta
BenchmarkExtractList/PodList-16             52806         8192          -84.49%
BenchmarkExtractList/ConfigMapList-16       22623         7978          -64.74%
BenchmarkEachListItem/PodList-16            3414          3335          -2.31%
BenchmarkEachListItem/ConfigMapList-16      3527          3473          -1.53%
BenchmarkReflectorList/PodList-16           1043821       978472        -6.26%
BenchmarkReflectorList/ConfigMapList-16     1044115       974130        -6.70%

benchmark                                   old allocs     new allocs     delta
BenchmarkExtractList/PodList-16             102            2              -98.04%
BenchmarkExtractList/ConfigMapList-16       102            2              -98.04%
BenchmarkEachListItem/PodList-16            1              1              +0.00%
BenchmarkEachListItem/ConfigMapList-16      1              1              +0.00%
BenchmarkReflectorList/PodList-16           4346           4246           -2.30%
BenchmarkReflectorList/ConfigMapList-16     4346           4245           -2.32%

benchmark                                   old bytes     new bytes     delta
BenchmarkExtractList/PodList-16             104200        1800          -98.27%
BenchmarkExtractList/ConfigMapList-16       30600         1800          -94.12%
BenchmarkEachListItem/PodList-16            8             8             +0.00%
BenchmarkEachListItem/ConfigMapList-16      8             8             +0.00%
BenchmarkReflectorList/PodList-16           1201093       1098691       -8.53%
BenchmarkReflectorList/ConfigMapList-16     1201082       1098545       -8.54%
```

Compare master and alloc variants.

> pkg: k8s.io/apimachinery/pkg/api/meta

```golang
benchcmp   master alloc                                                                                                          ubuntu@www
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                        old ns/op     new ns/op     delta
BenchmarkExtractListItem/FooList-16              183374        296696        +61.80%
BenchmarkExtractListItem/SampleList-16           214082        247796        +15.75%
BenchmarkExtractListItem/RawExtensionList-16     80363         86605         +7.77%
BenchmarkEachListItem/FooList-16                 34429         218988        +536.06%
BenchmarkEachListItem/SampleList-16              39539         222412        +462.51%
BenchmarkEachListItem/RawExtensionList-16        18509         121292        +555.31%

benchmark                                        old allocs     new allocs     delta
BenchmarkExtractListItem/FooList-16              1002           1002           +0.00%
BenchmarkExtractListItem/SampleList-16           1002           1002           +0.00%
BenchmarkExtractListItem/RawExtensionList-16     1002           1002           +0.00%
BenchmarkEachListItem/FooList-16                 1              1001           +100000.00%
BenchmarkEachListItem/SampleList-16              1              1001           +100000.00%
BenchmarkEachListItem/RawExtensionList-16        1              1001           +100000.00%

benchmark                                        old bytes     new bytes     delta
BenchmarkExtractListItem/FooList-16              304395        304394        -0.00%
BenchmarkExtractListItem/SampleList-16           304392        304393        +0.00%
BenchmarkExtractListItem/RawExtensionList-16     64392         64392         +0.00%
BenchmarkEachListItem/FooList-16                 8             288009        +3600012.50%
BenchmarkEachListItem/SampleList-16              8             288009        +3600012.50%
BenchmarkEachListItem/RawExtensionList-16        8             48008         +600000.00%
```

> pkg: k8s.io/client-go/tools/cache

```golang
benchcmp   master alloc                                                                                                         ubuntu@www
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                   old ns/op     new ns/op     delta
BenchmarkExtractList/PodList-16             52806         55915         +5.89%
BenchmarkExtractList/ConfigMapList-16       22623         27391         +21.08%
BenchmarkEachListItem/PodList-16            3414          52826         +1447.33%
BenchmarkEachListItem/ConfigMapList-16      3527          22740         +544.74%
BenchmarkReflectorList/PodList-16           1043821       1070159       +2.52%
BenchmarkReflectorList/ConfigMapList-16     1044115       1066411       +2.14%

benchmark                                   old allocs     new allocs     delta
BenchmarkExtractList/PodList-16             102            102            +0.00%
BenchmarkExtractList/ConfigMapList-16       102            102            +0.00%
BenchmarkEachListItem/PodList-16            1              101            +10000.00%
BenchmarkEachListItem/ConfigMapList-16      1              101            +10000.00%
BenchmarkReflectorList/PodList-16           4346           4346           +0.00%
BenchmarkReflectorList/ConfigMapList-16     4346           4346           +0.00%

benchmark                                   old bytes     new bytes     delta
BenchmarkExtractList/PodList-16             104200        104200        +0.00%
BenchmarkExtractList/ConfigMapList-16       30600         30600         +0.00%
BenchmarkEachListItem/PodList-16            8             102408        +1280000.00%
BenchmarkEachListItem/ConfigMapList-16      8             28808         +360000.00%
BenchmarkReflectorList/PodList-16           1201093       1201408       +0.03%
BenchmarkReflectorList/ConfigMapList-16     1201082       1201351       +0.02%
```